StatusBox::doMachineStatus: Extend and improve System Status - add os/architecture info - add system time - add Qt version - add network interface types and names, break out IP addresses for readability, add MAC address - make uptime, load and memory fields dynamic (i.e. they are updated at various intervals)

+++ mythtv/libs/libmyth/remoteutil.cpp
@@ -36,15 +36,15 @@ vector<ProgramInfo *> *RemoteGetRecordedList(int sort)
     return info;
 }
 
-bool RemoteGetLoad(float load[3])
+bool RemoteGetLoad(double load[3])
 {
     QStringList strlist(QString("QUERY_LOAD"));
 
     if (gCoreContext->SendReceiveStringList(strlist) && strlist.size() >= 3)
     {
-        load[0] = strlist[0].toFloat();
-        load[1] = strlist[1].toFloat();
-        load[2] = strlist[2].toFloat();
+        load[0] = strlist[0].toDouble();
+        load[1] = strlist[1].toDouble();
+        load[2] = strlist[2].toDouble();
         return true;
     }
