+++ mythtv/libs/libmyth/remoteutil.h
@@ -15,7 +15,7 @@ class ProgramInfo;
 class MythEvent;
 
 MPUBLIC vector<ProgramInfo *> *RemoteGetRecordedList(int sort);
-MPUBLIC bool RemoteGetLoad(float load[3]);
+MPUBLIC bool RemoteGetLoad(double load[3]);
 MPUBLIC bool RemoteGetUptime(time_t &uptime);
 MPUBLIC
 bool RemoteGetMemStats(int &totalMB, int &freeMB, int &totalVM, int &freeVM);
